refactor: feature oriented architecture for project-environment
