+++ src/lib/features/access/createAccessService.ts
@@ -3,14 +3,14 @@ import EventStore from '../../db/event-store';
 import GroupStore from '../../db/group-store';
 import { AccountStore } from '../../db/account-store';
 import RoleStore from '../../db/role-store';
-import EnvironmentStore from '../../db/environment-store';
+import EnvironmentStore from '../project-environment/environment-store';
 import { AccessStore } from '../../db/access-store';
 import { AccessService, EventService, GroupService } from '../../services';
 import FakeGroupStore from '../../../test/fixtures/fake-group-store';
 import FakeEventStore from '../../../test/fixtures/fake-event-store';
 import { FakeAccountStore } from '../../../test/fixtures/fake-account-store';
 import FakeRoleStore from '../../../test/fixtures/fake-role-store';
-import FakeEnvironmentStore from '../../../test/fixtures/fake-environment-store';
+import FakeEnvironmentStore from '../project-environment/fake-environment-store';
 import FakeAccessStore from '../../../test/fixtures/fake-access-store';
 import FeatureTagStore from '../../db/feature-tag-store';
 import FakeFeatureTagStore from '../../../test/fixtures/fake-feature-tag-store';
