+++ src/lib/features/feature-toggle/tests/feature-toggle-service.e2e.test.ts
@@ -10,7 +10,7 @@ import {
     IVariant,
     SKIP_CHANGE_REQUEST,
 } from '../../../types';
-import EnvironmentService from '../../../services/environment-service';
+import EnvironmentService from '../../project-environment/environment-service';
 import { ForbiddenError, PatternError, PermissionError } from '../../../error';
 import { ISegmentService } from '../../../segments/segment-service-interface';
 import { createFeatureToggleService, createSegmentService } from '../..';
