+++ src/lib/features/project/createProjectService.ts
@@ -2,7 +2,7 @@ import { Db, IUnleashConfig } from 'lib/server-impl';
 import EventStore from '../../db/event-store';
 import GroupStore from '../../db/group-store';
 import { AccountStore } from '../../db/account-store';
-import EnvironmentStore from '../../db/environment-store';
+import EnvironmentStore from '../project-environment/environment-store';
 import {
     AccessService,
     EventService,
@@ -28,7 +28,7 @@ import { FavoriteFeaturesStore } from '../../db/favorite-features-store';
 import { FavoriteProjectsStore } from '../../db/favorite-projects-store';
 import FakeProjectStore from '../../../test/fixtures/fake-project-store';
 import FakeFeatureToggleStore from '../feature-toggle/fakes/fake-feature-toggle-store';
-import FakeEnvironmentStore from '../../../test/fixtures/fake-environment-store';
+import FakeEnvironmentStore from '../project-environment/fake-environment-store';
 import FakeFeatureEnvironmentStore from '../../../test/fixtures/fake-feature-environment-store';
 import FakeProjectStatsStore from '../../../test/fixtures/fake-project-stats-store';
 import FakeFavoriteFeaturesStore from '../../../test/fixtures/fake-favorite-features-store';
