+++ src/test/e2e/stores/feature-environment-store.e2e.test.ts
@@ -3,7 +3,7 @@ import dbInit from '../helpers/database-init';
 import getLogger from '../../fixtures/no-logger';
 import { IFeatureEnvironmentStore } from '../../../lib/types/stores/feature-environment-store';
 import { IFeatureToggleStore } from '../../../lib/features/feature-toggle/types/feature-toggle-store-type';
-import { IEnvironmentStore } from '../../../lib/types/stores/environment-store';
+import { IEnvironmentStore } from '../../../lib/features/project-environment/environment-store-type';
 
 let db;
 let stores: IUnleashStores;
